requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pytest
httpx
passlib[bcrypt]
python-jose[cryptography]
email-validator
